improve performance + fix parse full screen width

+++ LED_Video_Unique_Cpp2/Color.h
@@ -1,15 +1,14 @@
 #pragma once
 #include <iostream>
-#include <string>
 
 using namespace std;
 
 namespace project {
 	struct Color
 	{
-		uint32_t r;
-		uint32_t g;
-		uint32_t b;
+		uint64_t r;
+		uint64_t g;
+		uint64_t b;
 	};
 }
